requirements.txt
annotated-types==0.7.0
click==8.2.1
confluent-kafka==2.11.0
iniconfig==2.1.0
packaging==25.0
pika==1.3.2
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
PyYAML==6.0.2
setuptools==80.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
